backend-services/services/ai-agent-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-talisman==1.1.0
PyJWT==2.8.0
cryptography==42.0.5
google-cloud-aiplatform>=1.38.0
google-cloud-logging
google-cloud-monitoring
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
google-cloud-bigtable
google-cloud-secret-manager
psycopg2-binary
redis
pytest==7.4.0
pytest-mock==3.11.1
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0
circuitbreaker
